+++ packages/node-fetch-server/src/lib/request-listener.ts
@@ -216,7 +216,17 @@ export async function sendResponse(
     }
   }
 
-  res.writeHead(response.status, headers)
+  if (res.req.httpVersionMajor === 1) {
+    res.writeHead(response.status, response.statusText, headers)
+  } else {
+    // HTTP/2 doesn't support status messages
+    // https://datatracker.ietf.org/doc/html/rfc7540#section-8.1.2.4
+    //
+    // HTTP2 `res.writeHead()` will safely ignore the statusText parameter, but
+    // it will emit a warning which we want to avoid.
+    // https://nodejs.org/docs/latest-v22.x/api/http2.html#responsewriteheadstatuscode-statusmessage-headers
+    res.writeHead(response.status, headers)
+  }
 
   if (response.body != null && res.req.method !== 'HEAD') {
     for await (let chunk of readStream(response.body)) {