+++ packages/node-fetch-server/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This is the changelog for [`node-fetch-server`](https://github.com/remix-run/remix/tree/main/packages/node-fetch-server). It follows [semantic versioning](https://semver.org/).
 
+## HEAD
+
+- Support `statusText` in HTTP/1 responses (#10745)
+
 ## v0.8.1 (2025-09-11)
 
 - Only abort `request.signal` when the connection closes before the response completes (see #10726)